Update Signal entries and add APK link
The titles and descriptions in the Signal entries were inconsistent with each other. It's also cleaner and more consistent with the other entries to only display one download link. I added a link to the Signal APK download site to the Related Information sections, because PrivacytoolsIO recommends operating systems that don't include Google Play, and the Signal APK download site isn't linked to directly from the signal.org front page.