Group and sort Firefox about:config tweaks
This makes it easier to apply the settings to Firefox, since you shouldn't have
to type in the full setting each time. I have applied these to probably a dozen
Firefox instances, and each time I have wished that the setting names were
sorted and grouped together.

I left the privacy settings at the top of the list, since they are the most
relevant.