Fix some typos, grammar, etc., and add details (#1418)
* Fix some typos, grammar, etc. on the site

Fixes some issues with typos, capitalization, grammar, and et cetera.

* Fix typo, grammar, etc. in repository

* Update README.md
Mention Discourse community earlier, add missing period

* Update CONTRIBUTING.md
i.e. is used for equivalence or clarification while e.g. is for
examples. For instance, we shouldn't say that IMAP is equivalent or an
explanation to all open-source software used to access email (e.g.
there's POP3, open-source clients to access when there isn't IMAP such
as Tutanota, etc.).

We also shouldn't call IMAP open-source software since it's a protocol.

* Change "socially motivated

* Apply suggestions from code review

Co-Authored-By: Jonah Aragon <jonah@triplebit.net>

* Suggestion from code review with extras

* 'Kill switch' to 'Killswitch"

* Consistency and minor additions to details

- More parallel sentence structures, following <Name> <Verb phrase>
for the first sentence of cards. Related to issue #1420.
- Make Njalla parallel to the others, and mention Njalla
is based in Nevis with VPS in Sweden
- Don't use "us" when talking about external services
- Orange Website also provides domain registration
- Update capitalization and add more hyphens
- Mention that TOS;DR evaluations are done by the community and that
  they also evaluate privacy policies (see https://edit.tosdr.org/about)
- "E2EE encryption" is redundant since "E2EE" already has "encryption"
in it. Might as well expand it since full term is used later on.
- <Name> <Verb phrase> structure for Magic Wormhole
- For consistency, don't start Worth Mentioning entries with the name
- https://english.stackexchange.com/questions/27707/post-hyphenation-of-split-compound-words
- All the other "alert alert-warning" don't repeat the generic name and 
we also don't say 'a software'

* instant-messenger: Remove <em>, more cleanup

* voice-video-messenger: Hyphens and cleanup

- We don't say "a software" so replace it with something that works

* paste-services: Cleanup & change cryptography info

As per CryptPad's whitepaper and FAQ:
https://cryptpad.fr/faq.html#security-crypto
https://blog.cryptpad.fr/images/CryptPad-Whitepaper-v1.0.pdf

Fixes #1417.

* encryption: "open-source"

* Page descriptions and other cleanup